Move load and save functions to MetsService
