+++ Kitodo/src/test/java/org/kitodo/production/services/dataformat/MetsServiceIT.java
@@ -14,8 +14,6 @@ package org.kitodo.production.services.dataformat;
 import static org.junit.Assert.assertEquals;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.InputStream;
 import java.net.URI;
 import java.util.Arrays;
 import java.util.stream.Collectors;
@@ -31,10 +29,8 @@ public class MetsServiceIT {
      */
     @Test
     public void testReadXML() throws Exception {
-        Workpiece workpiece;
-        try (InputStream in = new FileInputStream(new File("../Kitodo-DataFormat/src/test/resources/meta.xml"))) {
-            workpiece = MetsService.getInstance().load(in);
-        }
+        Workpiece workpiece = MetsService.getInstance()
+                .load(new File("../Kitodo-DataFormat/src/test/resources/meta.xml").toURI());
 
         // METS file has 183 associated images
         assertEquals(183, workpiece.getMediaUnits().size());
